fixed frames not displayed and using double instead of float for fps
Modified in PlayerLayoutManager::createAdvancedToolbar() "&AdvancedToolbar::updateFps" to &SimpleToolbar::updateFps, now fps are correctly set in advanced toolbar to display the current frame

+++ MediaWidget.h
@@ -39,7 +39,7 @@ signals:
     void updateSliderValueRequested(int64_t newTime);
     void activated(MediaWidget* self);
     void updateSliderRangeRequested(int64_t mediaDuration);
-    void updateFpsRequested(float fps);
+    void updateFpsRequested(double fps);
 };
 
 #endif // MEDIAWIDGET_H
